Refactor redux folder structure
Move movies reducers to a separate folder to match the correct convention
