--- app/src/redux/auth/types.ts
@@ -1,10 +1,0 @@
-import {
-    AuthenticationReducerType,
-    authenticationReducerName,
-} from '../auth/reducers/types';
-
-export const featuresReducerName = 'featureReducer';
-
-export interface FeaturesReducerType {
-    [authenticationReducerName]: AuthenticationReducerType;
-}
