+++ app/src/redux/movies/api/genres/index.ts
