+++ app/src/redux/movies/selectors/filterLimits/index.ts
@@ -1,8 +1,9 @@
-import { FilterLimitsSliceType } from 'redux/reducers/FilterLimits';
+import { FilterLimitsSliceType } from 'redux/movies/reducers/FilterLimits';
 import { createSelector } from '@reduxjs/toolkit';
+import { featuresReducerName } from 'redux/store';
 
 const getCoreState = (state: any): FilterLimitsSliceType =>
-    state.filterLimits as FilterLimitsSliceType;
+    state[featuresReducerName].movies.filterLimits as FilterLimitsSliceType;
 
 export const getFilterLimitsStatus = createSelector(
     [getCoreState],
