+++ app/src/redux/movies/selectors/movies/index.ts
@@ -1,8 +1,9 @@
-import { MovieSliceType } from 'redux/reducers/Movie';
+import { MovieSliceType } from 'redux/movies/reducers/Movie';
 import { createSelector } from '@reduxjs/toolkit';
+import { featuresReducerName } from 'redux/store';
 
 const getCoreState = (state: any): MovieSliceType =>
-    state.movies as MovieSliceType;
+    state[featuresReducerName].movies.movies as MovieSliceType;
 
 export const getMovies = createSelector(
     [getCoreState],
